requirements.txt
psycopg2-binary==2.9.9
# psycopg2==2.9.9
Flask==3.0.2
Flask-Cors==4.0.0
flask-restx==1.3.0
Werkzeug==3.0.1